third_party/blink/web_tests/fast/dom/shadow/sibling-rules-under-shadow-root-expected.txt

crbug.com/337573: Test direct/indirect adjucent and positinal rules under shadow root.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS colorOf(getNodeInComposedTree("host/child2")) is "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child2")) is "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child3")) is "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child1")) is "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child2")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child3")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child1")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child2")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child3")) is "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child1")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child2")) is "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child3")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child1")) is "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child2")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child3")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child1")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child2")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child3")) is "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child4")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child1")) is not "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child2")) is "rgb(0, 128, 0)"
PASS colorOf(getNodeInComposedTree("host/child3")) is not "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
